### 2024-11-19 #7505 @editorialbot
[PRE REVIEW]: `UnfoldMakie.jl`: EEG/ERP visualization package
**Submitting author:** @vladdez (<a href="http://orcid.org/0000-0002-4738-6655">Vladimir Mikheev</a>)
**Repository:** https://github.com/unfoldtoolbox/UnfoldMakie.jl
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v0.5.9
