### 2024-11-20 07e9dec
new test for butterfly with channel labels
