+++ test/test_butterfly.jl
@@ -186,6 +186,9 @@ end
 end =#
 
 @testset "butterfly: topoplot channel names" begin
-    plot_butterfly(df; positions = pos, 
-    topo_attributes = (; label_text = true, label_scatter = (; markersize = 10)))
+    plot_butterfly(
+        df;
+        positions = pos,
+        topo_attributes = (; label_text = true, label_scatter = (; markersize = 10)),
+    )
 end
