+++ previews/PR278/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"generated/how_to/position2color/","page":"Change colormap of Butterfly plot","title":"Change colormap of Butterfly plot","text":"EditURL = \"../../../literate/how_to/position2color.jl\"","category":"page"},{"location":"generated/how_to/position2color/#pos2color","page":"Change colormap of Butterfly plot","title":"Change colormap of Butterfly plot ","text":"","category":"section"},{"location":"generated/how_to/position2color/","page":"Change colormap of Butterfly plot","title":"Change colormap of Butterfly plot","text":"You want to change the colors of the lines and markers on the inserted topoplot. To do that you need to change the color scheme (aka color map) of the butterfly plot. You can find th elist of colormaps for Makie here.","category":"page"},{"location":"generated/how_to/position2color/#Setup","page":"Change colormap of Butterfly plot","title":"Setup","text":"","category":"section"},{"location":"generated/how_to/position2color/","page":"Change colormap of Butterfly plot","title":"Change colormap of Butterfly plot","text":"using UnfoldMakie\nusing CairoMakie\nusing DataFramesMeta\nusing Colors","category":"page"},{"location":"generated/how_to/position2color/","page":"Change colormap of Butterfly plot","title":"Change colormap of Butterfly plot","text":"By default the plot looks like this:","category":"page"},{"location":"generated/how_to/position2color/","page":"Change colormap of Butterfly plot","title":"Change colormap of Butterfly plot","text":"include(\"../../../example_data.jl\")\nresults, positions = example_data(\"TopoPlots.jl\")\nplot_butterfly(results; positions = positions)","category":"page"},{"location":"generated/how_to/position2color/#Color-schemes","page":"Change colormap of Butterfly plot","title":"Color schemes","text":"","category":"section"},{"location":"generated/how_to/position2color/#MNE-style","page":"Change colormap of Butterfly plot","title":"MNE style","text":"","category":"section"},{"location":"generated/how_to/position2color/","page":"Change colormap of Butterfly plot","title":"Change colormap of Butterfly plot","text":"We can change the color scale by specifying a function that maps from an (x, y) tuple to a color. UnfoldMakie currently provides three different color scales:","category":"page"},{"location":"generated/how_to/position2color/","page":"Change colormap of Butterfly plot","title":"Change colormap of Butterfly plot","text":"pos2colorRGB (same as MNE-Python),\npos2colorHSV (HSV color space),\npos2colorRomaO.","category":"page"},{"location":"generated/how_to/position2color/","page":"Change colormap of Butterfly plot","title":"Change colormap of Butterfly plot","text":"While RGB & HSV have the advantage of being 2D color maps, Roma0 has the advantage of being perceptually uniform. Also you can specify a uniform color.","category":"page"},{"location":"generated/how_to/position2color/","page":"Change colormap of Butterfly plot","title":"Change colormap of Butterfly plot","text":"plot_butterfly(\n    results;\n    positions = positions,\n    topopositions_to_color = pos -> UnfoldMakie.pos_to_color_RGB(pos),\n)","category":"page"},{"location":"generated/how_to/position2color/#HSV-Space","page":"Change colormap of Butterfly plot","title":"HSV-Space","text":"","category":"section"},{"location":"generated/how_to/position2color/","page":"Change colormap of Butterfly plot","title":"Change colormap of Butterfly plot","text":"plot_butterfly(\n    results;\n    positions = position
[…]
