### 2024-12-03 #7505 @Kevin-Mattheus-Moerman
@vladdez Dear author, thanks for this submission. I am the AEiC on this track and here to help process the initial steps. I've just assigned a handling editor. However, before we proceed, please can you have a look at the following points:
