+++ joss_paper/paper.bib
@@ -33,6 +33,7 @@
   author={Jung, Tzyy-Ping and Makeig, Scott and Westerfield, Marissa and Townsend, Jeanne and Courchesne, Eric and Sejnowski, Terrence J},
   journal={Advances in neural information processing systems},
   volume={11},
+  doi={10.1002/hbm.1050},
   year={1998}
 }
 
@@ -50,6 +51,7 @@
 @article{ehinger2018eegvis,
   title={EEGVIS toolbox},
   author={Ehinger, Benedikt V},
+  doi={10.5281/zenodo.1312813},
   journal={Osnabr{\"u}ck.[Google Scholar]},
   year={2018}
 }
@@ -93,6 +95,7 @@
   volume={13},
   pages={76},
   year={2019},
+  doi={10.3389/fnins.2019.00076},
   publisher={Frontiers Media SA}
 }
 
@@ -158,6 +161,7 @@
   title =	 "{Robust benchmarking in noisy environments}",
   journal =	 {arXiv e-prints},
   keywords =	 {Computer Science - Performance, 68N30, B.8.1, D.2.5},
+  doi = {10.48550/arXiv.1608.04295},
   year =	 2016,
   month =	 "Aug",
   eid =		 {arXiv:1608.04295},
@@ -189,7 +193,8 @@
     year={2024},
     author = {Ehinger, Benedikt V and {Phillip} Alday and {Singhvi}, Anshul and {Danisch} Simon},
     title = {TopoPlots.jl: A Julia package for topoplot generation},
-    version = {0.1.9},
+    version = {0.2.0},
+    doi={10.5281/zenodo.14179492},
     url = {https://github.com/MakieOrg/TopoPlots.jl}
   }
 
@@ -198,6 +203,7 @@
     author = {Phillip Alday and Alex Arslan},
     title = {PyMNE.jl},
     version = {0.2.2},
+    doi={10.5281/zenodo.14172218},
     url = {https://github.com/beacon-biosignals/PyMNE.jl}
   }
 
@@ -275,5 +281,6 @@ author = {Simon Kornblith and
   author = {Schepers, Judith and Lips, Luis and Marathe, Maanik and Ehinger, Benedikt},
   title = {{UnfoldSim.jl}},
   year={2024},
+  doi={10.5281/zenodo.13933412},
   url = {https://github.com/unfoldtoolbox/UnfoldSim.jl}
 }
