+++ joss_paper/paper.bib
@@ -132,7 +132,7 @@
   doi = {10.3389/fnins.2013.00267},
   journal = {Frontiers in Neuroscience},
   number = {267},
-  pages = {1--13},
+  pages = {1-13},
   title = {{MEG and EEG Data Analysis with MNE-Python}},
   volume = {7},
   year = {2013}
