+++ joss_paper/paper.bib
@@ -1,0 +1,286 @@
+@article{mikheev2024art,
+  title={The Art of Brainwaves: A Survey on Event-Related Potential Visualization Practices},
+  author={Mikheev, Vladimir and Skukies, Rene and Ehinger, Benedikt V},
+  journal={Aperture Neuro},
+  doi = {10.52294/001c.116386},
+  volume={4},
+  year={2024},
+  publisher={Organization for Human Brain Mapping}
+}
+
+@article{ehinger2019unfold,
+  title={Unfold: an integrated toolbox for overlap correction, non-linear modeling, and regression-based EEG analysis},
+  author={Ehinger, Benedikt V and Dimigen, Olaf},
+  journal={PeerJ},
+  doi={10.5281/zenodo.5759066},
+  volume={7},
+  pages={e7838},
+  year={2019},
+  publisher={PeerJ Inc.}
+}
+@article{danisch2021makie,
+  title={Makie.jl: Flexible high-performance data visualization for Julia},
+  author={Danisch, Simon and Krumbiegel, Julius},
+  journal={Journal of Open Source Software},
+  volume={6},
+  number={65},
+  pages={3349},
+  year={2021}
+}
+
+@article{jung1998analyzing,
+  title={Analyzing and visualizing single-trial event-related potentials},
+  author={Jung, Tzyy-Ping and Makeig, Scott and Westerfield, Marissa and Townsend, Jeanne and Courchesne, Eric and Sejnowski, Terrence J},
+  journal={Advances in neural information processing systems},
+  volume={11},
+  doi={10.1002/hbm.1050},
+  year={1998}
+}
+
+@article{ten2006design,
+  title={Design and evaluation of tiled parallel coordinate visualization of multichannel EEG data},
+  author={Ten Caat, Michael and Maurits, Natasha M and Roerdink, Jos BTM},
+  journal={IEEE Transactions on Visualization and Computer Graphics},
+  volume={13},
+  number={1},
+  pages={70--79},
+  year={2006},
+  publisher={IEEE}
+}
+
+@article{ehinger2018eegvis,
+  title={EEGVIS toolbox},
+  author={Ehinger, Benedikt V},
+  doi={10.5281/zenodo.1312813},
+  journal={Osnabr{\"u}ck.[Google Scholar]},
+  year={2018}
+}
+
+@article{delorme2004eeglab,
+  title={EEGLAB: an open source toolbox for analysis of single-trial EEG dynamics including independent component analysis},
+  author={Delorme, Arnaud and Makeig, Scott},
+  journal={Journal of neuroscience methods},
+  volume={134},
+  number={1},
+  pages={9--21},
+  year={2004},
+  publisher={Elsevier}
+}
+
+@article{oostenveld2011fieldtrip,
+  title={FieldTrip: open source software for advanced analysis of MEG, EEG, and invasive electrophysiological data},
+  author={Oostenveld, Robert and Fries, Pascal and Maris, Eric and Schoffelen, Jan-Mathijs},
+  journal={Computational intelligence and neuroscience},
+  volume={2011},
+  number={1},
+  pages={156869},
+  year={2011},
+  publisher={Wiley Online Library}
+}
+
+@article{lopez2014erplab,
+  title={ERPLAB: an open-source toolbox for the analysis of event-related potentials},
+  author={Lopez-Calderon, Javier and Luck, Steven J},
[…]
