### 2024-12-03 #7505 @fabian-s
hi @mscaudill @AdamWysokinski @rob-luke @ferchaure :wave:
Would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
