+++ joss_paper/paper_UM.md
@@ -30,7 +30,7 @@ affiliations:
    index: 2
 
 date: 30 October 2024
-bibliography: paper_cit.bib
+bibliography: paper_UM.bib
 
 ---
 
