+++ joss_paper/paper_UM.bib
@@ -22,6 +22,7 @@
   title={Makie.jl: Flexible high-performance data visualization for Julia},
   author={Danisch, Simon and Krumbiegel, Julius},
   journal={Journal of Open Source Software},
+  doi={10.21105/joss.03349},
   volume={6},
   number={65},
   pages={3349},
@@ -43,7 +44,8 @@
   journal={IEEE Transactions on Visualization and Computer Graphics},
   volume={13},
   number={1},
-  pages={70--79},
+  doi={10.1109/tvcg.2007.9},
+  pages={70-79},
   year={2006},
   publisher={IEEE}
 }
@@ -54,15 +56,16 @@
   doi={10.5281/zenodo.1312813},
   journal={Osnabr{\"u}ck.[Google Scholar]},
   year={2018}
-}
+}F
 
 @article{delorme2004eeglab,
   title={EEGLAB: an open source toolbox for analysis of single-trial EEG dynamics including independent component analysis},
   author={Delorme, Arnaud and Makeig, Scott},
   journal={Journal of neuroscience methods},
   volume={134},
+  doi={10.1016/j.jneumeth.2003.10.009},
   number={1},
-  pages={9--21},
+  pages={9-21},
   year={2004},
   publisher={Elsevier}
 }
@@ -71,6 +74,7 @@
   title={FieldTrip: open source software for advanced analysis of MEG, EEG, and invasive electrophysiological data},
   author={Oostenveld, Robert and Fries, Pascal and Maris, Eric and Schoffelen, Jan-Mathijs},
   journal={Computational intelligence and neuroscience},
+  doi={10.1155/2011/156869},
   volume={2011},
   number={1},
   pages={156869},
@@ -82,6 +86,7 @@
   title={ERPLAB: an open-source toolbox for the analysis of event-related potentials},
   author={Lopez-Calderon, Javier and Luck, Steven J},
   journal={Frontiers in human neuroscience},
+  doi={10.3389/fnhum.2014.00213},
   volume={8},
   pages={213},
   year={2014},
@@ -175,6 +180,7 @@
   title={EEGVIS: a Matlab toolbox for browsing, exploring, and viewing large datasets},
   author={Robbins, Kay A},
   journal={Frontiers in Neuroinformatics},
+  doi={10.3389/fninf.2012.00017},
   volume={6},
   pages={17},
   year={2012},
@@ -212,6 +218,7 @@
   title={The misuse of colour in science communication},
   author={Crameri, Fabio and Shephard, Grace E and Heron, Philip J},
   journal={Nature communications},
+  doi={10.1038/s41467-020-19160-7},
   volume={11},
   number={1},
   pages={5444},
@@ -223,9 +230,9 @@
   title={Why We Use Bad Color Maps and What You Can Do About It.},
   author={Moreland, Kenneth D},
   year={2015},
+  doi={10.2352/issn.2470-1173.2016.16.hvei-133},
   institution={Sandia National Lab.(SNL-NM), Albuquerque, NM (United States)}
 }
-
 
 @software{simon_kornblith_2023_8344531,
 author = {Simon Kornblith and
