### 2024-12-03 #7505 @mscaudill
Hi Fabian,

Thank you for inviting me to review this submission. I am not familiar with
Julia so I will need to decline.

Best
Matt Caudill

On Tue, Dec 3, 2024 at 7:52 AM Fabian Scheipl ***@***.***>
wrote:

> hi @mscaudill <https://github.com/mscaudill> @AdamWysokinski
> <https://github.com/AdamWysokinski> @rob-luke
> <https://github.com/rob-luke> @ferchaure <https://github.com/ferchaure> 👋
> Would you be willing to review this submission for JOSS? We carry out our
> checklist-driven reviews here in GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
