### 2024-12-03 #7505 @AdamWysokinski
Count me in.
