### 2024-12-03 #7560 @editorialbot
[REVIEW]: `UnfoldMakie.jl`: EEG/ERP visualization package
**Submitting author:** @vladdez (<a href="http://orcid.org/0000-0002-4738-6655">Vladimir Mikheev</a>)
**Repository:** https://github.com/unfoldtoolbox/UnfoldMakie.jl
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v0.5.11.post1

Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue in the <a href="https://github.com/unfoldtoolbox/UnfoldMakie.jl " target="_blank">target repository</a> and link to those issues (especially acceptance-blockers) by leaving comments in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice versa will create corresponding breadcrumb trails in the link target.)
