+++ Project.toml
@@ -1,7 +1,7 @@
 name = "UnfoldMakie"
 uuid = "69a5ce3b-64fb-4f22-ae69-36dd4416af2a"
 authors = ["Vladimir Mikheev", "Daniel Baumgartner", "Sören Döring", "Niklas Gärtner", "Furkan Lokman", "Benedikt Ehinger"]
-version = "0.5.9"
+version = "0.5.10"
 
 [deps]
 AlgebraOfGraphics = "cbdf2221-f076-402e-a563-3d30da359d67"
@@ -44,7 +44,7 @@ CoordinateTransformations = "0.6"
 DataFrames = "1"
 DataStructures = "0.18"
 DocStringExtensions = "0.9"
-GeometryBasics = "0.4"
+GeometryBasics = "0.4, 0.5"
 GridLayoutBase = "0.9, 0.10, 0.11"
 ImageFiltering = "0.7"
 Interpolations = "0.14, 0.15"
@@ -55,7 +55,7 @@ PyMNE = "0.2"
 SparseArrays = "1"
 StaticArrays = "1"
 Statistics = "1"
-TopoPlots = "0.1"
+TopoPlots = "0.1, 0.2"
 Unfold = "0.6, 0.7"
 julia = "1"
 
