+++ README.md
@@ -8,7 +8,7 @@
 [![Dev](https://img.shields.io/badge/docs-dev-blue.svg)](https://unfoldtoolbox.github.io/UnfoldMakie.jl/dev)
 [![Build Status](https://github.com/unfoldtoolbox/UnfoldMakie.jl/workflows/CI/badge.svg)](https://github.com/unfoldtoolbox/UnfoldMakie.jl/actions)
 [![Coverage](https://codecov.io/gh/behinger/UnfoldMakie.jl/branch/master/graph/badge.svg)](https://codecov.io/gh/behinger/UnfoldMakie.jl)
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10235220.svg)](https://doi.org/10.5281/zenodo.10235220)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.14192333.svg)](https://doi.org/10.5281/zenodo.14192333)
 
 
 
