### 2024-12-16 #7560 @vladdez
v0.5.11
[10.5281/zenodo.14500860](https://doi.org/10.5281/zenodo.14500860)
