### 2024-12-17 #7560 @fabian-s
@vladdez the deposit in zenodo is not the `joss-paper` branch though, is it? it should be...
suggest to merge https://github.com/unfoldtoolbox/UnfoldMakie.jl/pull/266 first, then archive the merged `main` branch.
