+++ CITATION.cff
@@ -20,7 +20,7 @@ authors:
   orcid: 'https://orcid.org/0000-0002-6276-3332'
 identifiers:
   - type: doi
-    value: 10.5281/zenodo.14046394
+    value: 10.5281/zenodo.14192333
 repository-code: 'https://github.com/unfoldtoolbox/UnfoldMakie.jl'
-version: 0.5.9
+version: 0.5.10
 date-released: '2023-10-24'
