+++ .github/workflows/joss_gen.yml
@@ -26,7 +26,7 @@ jobs:
           fetch-depth: 0
         #uses: actions/checkout@v3.5.3
       - name: Checkout latest commit
-        run: git pull && GITHUB_SHA=$(git rev-parse HEAD)
+        run: GITHUB_SHA=$(git rev-parse HEAD)
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
         with:
