### 2024-12-17 #7560 @fabian-s
@vladdez please see my review comments in your merged PR -- sorry, meant to give you those before I asked you to merge the PR but did not click the final submit button... :(
