### 2024-12-18 #7560 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
