### 2024-12-19 79ba9ea
refactoring topo_series: Bene makes a first strike
