+++ src/eeg_series.jl
@@ -34,46 +34,26 @@ function eeg_array_to_dataframe(data::AbstractMatrix, label_aliases::AbstractVec
 end
 
 """
-    eeg_topoplot_series(data::DataFrame,
-        f,
-        data::DataFrame;
-        bin_width = nothing,
-        bin_num = nothing,
-        y = :erp,
-        label = :label,
-        col = :time,
-        row = nothing,
-        col_labels = false,
-        row_labels = false,
+     eeg_topoplot_series(data::DataFrame,
+        fig,
+        data_inp::Union{<:Observable,<:AbstractMatrix};
+        layout = nothing,
+        xlabels = nothing, # can be a vector too
+        labels = nothing,
         rasterize_heatmaps = true,
-        combinefun = mean,
-        xlim_topo,
-        ylim_topo,
-        topo_attributes...,
+        interactive_scatter = nothing,
+        highlight_scatter = false,
+        topo_axis = (;),
+        topo_attributes = (;),
+        positions,
     )
     eeg_topoplot_series!(fig, data::DataFrame; kwargs..)
 
 Plot a series of topoplots. 
 The function takes the `combinefun = mean` over the `:time` column of `data`.
-- `f` \\
+- `fig` \\
     Figure object. \\
-- `data::DataFrame`\\
-    Needs the columns `:time` and `y(=:erp)`, and `label(=:label)`. \\
-    If `data` is a matrix, it is automatically cast to a dataframe, time bins are in samples, labels are `string.(1:size(data,1))`.
-- `bin_width::Real = nothing`\\
-    Number specifing the width of bin of continuous x-value in its units.\\
-- `bin_num::Real = nothing`\\
-    Number of topoplots.\\
-    Either `bin_width`, or `bin_num` should be specified. Error if they are both specified\\
-    If `mapping.col` or `mapping.row` are categorical `bin_width` and `bin_num` stay as `nothing`.
-- `col`, `row = :time` \\
-    Specify the field to be divided into columns and rows. The default is `col = :time` to split by the time field and `row = nothing`. \\
-    Useful to split by a condition, e.g. `...(..., col = :time, row = :condition)` would result in multiple (as many as different values in `df.condition`) rows of topoplot series.
-- `row_labels`, `col_labels = false` \\
-    Indicate whether there should be labels in the plots in the first column to indicate the row value and in the last row to indicate the time (typically timerange).
-- `combinefun::Function = mean`\\
-    Specify how the samples are summarised.\\
-    Example functions: `mean`, `median`, `std`.  
+- `data::Union{<:Observable,<:AbstractMatrix}`\\
 - `topo_axis::NamedTuple = (;)`\\
     Here you can flexibly change configurations of the topoplot axis.\\
     To see all options just type `?Axis` in REPL.\\
@@ -93,46 +73,12 @@ function eeg_topoplot_series(
     return eeg_topoplot_series!(Figure(; figure...), data; kwargs...)
 end
 
-# AbstractMatrix
 function eeg_topoplot_series!(
     fig,
-    data::Union{<:Observable,<:DataFrame,<:AbstractMatrix};
-    kwargs...,
-)
-    return eeg_topoplot_series!(fig, data, string.(1:size(data, 1)); kwargs...)
-end
-
-# convert a 2D Matrix to the dataframe
-function eeg_topoplot_series(
-    data::Union{<:Observable,<:DataFrame,<:AbstractMatrix},
-    labels;
-    kwargs...,
-)
-    return eeg_topoplot_series(eeg_array_to_dataframe(data, labels); kwargs...)
-end
-function eeg_topoplot_series!(
-    fig,
[…]
