+++ src/plot_topoplotseries.jl
@@ -59,7 +59,7 @@ plot_topoplotseries(data::Union{<:Observable{<:DataFrame},DataFrame}; kwargs...)
 
 function plot_topoplotseries!(
     f::Union{GridPosition,GridLayout,Figure,GridLayoutBase.GridSubposition},
-    data::Union{<:Observable{<:DataFrame},DataFrame};
+    data_inp::Union{<:Observable{<:DataFrame},DataFrame};
     bin_width = nothing,
     bin_num = nothing,
     positions = nothing,
@@ -75,8 +75,8 @@ function plot_topoplotseries!(
     kwargs...,
 )
 
-    data = _as_observable(data)
-    data_cuts = data
+    data = _as_observable(data_inp)
+    data_cuts = @lift deepcopy($data)
     positions = get_topo_positions(; positions = positions, labels = labels)
     chan_or_label = "label" ∉ names(to_value(data)) ? :channel : :label
 
@@ -96,33 +96,44 @@ function plot_topoplotseries!(
     end
     # resolve columns with data
     config.mapping = resolve_mappings(to_value(data), config.mapping)
-    data_copy = deepcopy(to_value(data)) # deepcopy prevents overwriting initial data
+
+    #data_copy = deepcopy(to_value(data)) # deepcopy prevents overwriting initial data
     cat_or_cont_columns =
-        eltype(data_copy[!, config.mapping.col]) <: Number ? "cont" : "cat"
-    if cat_or_cont_columns == "cat"
+        @lift eltype($data[!, config.mapping.col]) <: Number ? "cont" : "cat"
+    if to_value(cat_or_cont_columns) == "cat"
         # overwrite 'Time windows [s]' default if categorical
         n_topoplots =
-            number_of_topoplots(data_copy; bin_width, bin_num, bins = 0, config.mapping)
-        ix =
-            findall.(
-                isequal.(unique(data_copy[!, config.mapping.col])),
-                [data_copy[!, config.mapping.col]],
-            )
+            number_of_topoplots(data; bin_width, bin_num, bins = 0, config.mapping)
+        #ix =
+        #    findall.(
+        #        isequal.(unique(data_copy[!, config.mapping.col])),
+        #        [data_copy[!, config.mapping.col]],
+        #    )
     else
-        bins = bins_estimation(
-            data_copy[!, config.mapping.col];
+        bins = @lift bins_estimation(
+            $data[!, config.mapping.col];
             bin_width,
             bin_num,
-            cat_or_cont_columns,
+            cat_or_cont_columns = $cat_or_cont_columns,
         )
-        n_topoplots =
-            number_of_topoplots(data_copy; bin_width, bin_num, bins, config.mapping)
-        to_value(data_cuts).cont_cuts =
-            cut(to_value(data_cuts)[!, config.mapping.col], bins; extend = true)
-        unique_cuts = unique(to_value(data_cuts).cont_cuts)
-        ix = findall.(isequal.(unique_cuts), [to_value(data).cont_cuts])
+        n_topoplots = @lift number_of_topoplots(
+            $data;
+            bin_width,
+            bin_num,
+            bins = $bins,
+            config.mapping,
+        )
+
+        cont_cuts = @lift cut($data[!, config.mapping.col], $bins; extend = true)
+
+        #unique_cuts = unique(to_value(data_cuts).cont_cuts)
+
+        #ix = findall.(isequal.(unique_cuts), [to_value(data_cuts).cont_cuts])
     end
-    data_row = @lift row_col_management($data_cuts, ix, n_topoplots, nrows, config)
[…]
