### 2024-12-19 e82262a
upd docstrings
