### 2024-12-19 8a150ae
upd docstrings
