### 2024-12-19 483482c
upd docstrings
