### 2024-12-20 #7560 @fabian-s
@vladdez we wait for @openjournals/bcm-eics to accept and process the Submission....
