### 2025-01-08 #7560 @fabian-s
@openjournals/bcm-eics
this is ready to accept from my side, please proceed
