### 2025-01-10 #7560 @vladdez
I have only one question: is it okay that the title has rendered as <code>UnfoldMakie.jl</code>: EEG/ERP visualization package.
Is it possible to get rid of "code"?
