### 2025-01-10 #7560 @vladdez
Thanks!
Will it take time to update? Because currently I do not see the difference.
Will the doi stay the same or will change?
Sorry for the inconvenience.
![image](https://github.com/user-attachments/assets/1b242b5e-1e69-44fe-8cdf-ae7c658bb27f)
