### 2025-01-10 #7560 @Kevin-Mattheus-Moerman
@vladdez yes it could take a bit of time but usually minutes rather than this long. I think it may still build the paper from the `joss_paper` branch. Is that still accurate or should we use the main branch now?
